@@ -8,28 +8,24 @@ import com.legobmw99.allomancy.modules.powers.data.AllomancerCapability;
 import com.legobmw99.allomancy.modules.powers.data.AllomancerDataProvider;
 import com.legobmw99.allomancy.modules.powers.network.UpdateEnhancedPacket;
 import com.legobmw99.allomancy.network.Network;
-import net.minecraft.block.BlockState;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.MobEntity;
-import net.minecraft.entity.item.ItemEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.player.ServerPlayerEntity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.particles.ParticleTypes;
-import net.minecraft.pathfinding.Path;
-import net.minecraft.pathfinding.PathPoint;
-import net.minecraft.potion.EffectInstance;
-import net.minecraft.potion.Effects;
-import net.minecraft.tileentity.ITickableTileEntity;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.DamageSource;
-import net.minecraft.util.RegistryKey;
-import net.minecraft.util.math.AxisAlignedBB;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.GameRules;
-import net.minecraft.world.World;
-import net.minecraft.world.server.ServerWorld;
+import net.minecraft.core.BlockPos;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.item.ItemEntity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.GameRules;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.TickingBlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.AABB;
 import net.minecraftforge.event.AttachCapabilitiesEvent;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.event.entity.living.LivingDeathEvent;
@@ -39,25 +35,21 @@ import net.minecraftforge.event.entity.player.PlayerSetSpawnEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
 import java.util.Arrays;
-import java.util.List;
-import java.util.Objects;
 
 public class CommonEventHandler {
 
     @SubscribeEvent
     public static void onAttachCapability(final AttachCapabilitiesEvent<Entity> event) {
-        if (event.getObject() instanceof PlayerEntity) {
+        if (event.getObject() instanceof Player) {
             AllomancerDataProvider provider = new AllomancerDataProvider();
             event.addCapability(AllomancerCapability.IDENTIFIER, provider);
-            event.addListener(provider::invalidate);
         }
     }
 
     @SubscribeEvent
     public static void onJoinWorld(final PlayerEvent.PlayerLoggedInEvent event) {
         if (!event.getPlayer().level.isClientSide) {
-            if (event.getPlayer() instanceof ServerPlayerEntity) {
-                ServerPlayerEntity player = (ServerPlayerEntity) event.getPlayer();
+            if (event.getPlayer() instanceof ServerPlayer player) {
 
                 player.getCapability(AllomancerCapability.PLAYER_CAP).ifPresent(data -> {
                     //Handle random misting case
@@ -66,7 +58,7 @@ public class CommonEventHandler {
                         data.addPower(Metal.getMetal(randomMisting));
                         ItemStack flakes = new ItemStack(MaterialsSetup.FLAKES.get(randomMisting).get());
                         // Give the player one flake of their metal
-                        if (!player.inventory.add(flakes)) {
+                        if (!player.getInventory().add(flakes)) {
                             ItemEntity entity = new ItemEntity(player.getCommandSenderWorld(), player.position().x(), player.position().y(), player.position().z(), flakes);
                             player.getCommandSenderWorld().addFreshEntity(entity);
                         }
@@ -83,9 +75,11 @@ public class CommonEventHandler {
     public static void onPlayerClone(final PlayerEvent.Clone event) {
         if (!event.getPlayer().level.isClientSide()) {
 
-            PlayerEntity player = event.getPlayer();
-            player.getCapability(AllomancerCapability.PLAYER_CAP).ifPresent(data -> {
+            event.getOriginal().reviveCaps();
 
+            Player player = event.getPlayer();
+
+            player.getCapability(AllomancerCapability.PLAYER_CAP).ifPresent(data -> {
                 event.getOriginal().getCapability(AllomancerCapability.PLAYER_CAP).ifPresent(oldData -> {
                     data.setDeathLoc(oldData.getDeathLoc(), oldData.getDeathDim());
                     if (!oldData.isUninvested()) { // make sure the new player has the same power status
@@ -104,6 +98,8 @@ public class CommonEventHandler {
                     }
                 });
             });
+            event.getOriginal().getCapability(AllomancerCapability.PLAYER_CAP).invalidate();
+            event.getOriginal().invalidateCaps();
 
             Network.sync(player);
         }
@@ -126,8 +122,7 @@ public class CommonEventHandler {
     @SubscribeEvent
     public static void onStartTracking(final net.minecraftforge.event.entity.player.PlayerEvent.StartTracking event) {
         if (!event.getTarget().level.isClientSide) {
-            if (event.getTarget() instanceof ServerPlayerEntity) {
-                ServerPlayerEntity player = (ServerPlayerEntity) event.getTarget();
+            if (event.getTarget() instanceof ServerPlayer player) {
                 Network.sync(player);
             }
         }
@@ -135,8 +130,7 @@ public class CommonEventHandler {
 
     @SubscribeEvent
     public static void onSetSpawn(final PlayerSetSpawnEvent event) {
-        PlayerEntity player = event.getPlayer();
-        if (player instanceof ServerPlayerEntity) {
+        if (event.getPlayer() instanceof ServerPlayer player) {
             player.getCapability(AllomancerCapability.PLAYER_CAP).ifPresent(data -> {
                 data.setSpawnLoc(event.getNewSpawn(), event.getSpawnWorld());
                 Network.sync(data, player);
@@ -146,8 +140,7 @@ public class CommonEventHandler {
 
     @SubscribeEvent
     public static void onLivingDeath(final LivingDeathEvent event) {
-        if (event.getEntityLiving() instanceof ServerPlayerEntity) {
-            ServerPlayerEntity player = (ServerPlayerEntity) event.getEntityLiving();
+        if (event.getEntityLiving() instanceof ServerPlayer player) {
             player.getCapability(AllomancerCapability.PLAYER_CAP).ifPresent(data -> {
                 data.setDeathLoc(new BlockPos(player.position()), player.level.dimension());
                 Network.sync(data, player);
@@ -158,8 +151,7 @@ public class CommonEventHandler {
     @SubscribeEvent
     public static void onDamage(final LivingHurtEvent event) {
         // Increase outgoing damage for pewter burners
-        if (event.getSource().getEntity() instanceof ServerPlayerEntity) {
-            ServerPlayerEntity source = (ServerPlayerEntity) event.getSource().getEntity();
+        if (event.getSource().getEntity() instanceof ServerPlayer source) {
             source.getCapability(AllomancerCapability.PLAYER_CAP).ifPresent(data -> {
 
                 if (data.isBurning(Metal.PEWTER)) {
@@ -176,16 +168,16 @@ public class CommonEventHandler {
                 }
 
                 if (data.isBurning(Metal.CHROMIUM)) {
-                    if (event.getEntityLiving() instanceof PlayerEntity) {
-                        PowerUtils.wipePlayer((PlayerEntity) event.getEntityLiving());
+                    if (event.getEntityLiving() instanceof Player player) {
+                        PowerUtils.wipePlayer(player);
                     }
                 }
             });
         }
 
         // Reduce incoming damage for pewter burners
-        if (event.getEntityLiving() instanceof ServerPlayerEntity) {
-            event.getEntityLiving().getCapability(AllomancerCapability.PLAYER_CAP).ifPresent(data -> {
+        if (event.getEntityLiving() instanceof ServerPlayer player) {
+            player.getCapability(AllomancerCapability.PLAYER_CAP).ifPresent(data -> {
                 if (data.isBurning(Metal.PEWTER)) {
                     if (data.isEnhanced()) { // Duralumin invulnerability
                         Allomancy.LOGGER.debug("Canceling Damage");
@@ -197,7 +189,7 @@ public class CommonEventHandler {
                         event.setAmount(event.getAmount() - 2);
                         // Note that they took damage, will come in to play if they stop burning
                         data.setDamageStored(data.getDamageStored() + 1);
-                        Network.sync(data, (ServerPlayerEntity) event.getEntityLiving());
+                        Network.sync(data, player);
                     }
                 }
             });
@@ -208,10 +200,10 @@ public class CommonEventHandler {
     public static void onWorldTick(final TickEvent.WorldTickEvent event) {
         if (event.phase == TickEvent.Phase.END) {
 
-            World world = event.world;
-            List<? extends PlayerEntity> list = world.players();
+            Level level = event.world;
+            var list = level.players();
             for (int enti = list.size() - 1; enti >= 0; enti--) {
-                PlayerEntity curPlayer = list.get(enti);
+                Player curPlayer = list.get(enti);
                 curPlayer.getCapability(AllomancerCapability.PLAYER_CAP).ifPresent(data -> {
                     if (!data.isUninvested()) {
 
@@ -235,8 +227,8 @@ public class CommonEventHandler {
 
                         // Run the necessary updates on the player's metals
                         // Ran AFTER duralumin and aluminum to make sure they function correctly
-                        if (curPlayer instanceof ServerPlayerEntity) {
-                            data.tickBurning((ServerPlayerEntity) curPlayer);
+                        if (curPlayer instanceof ServerPlayer player) {
+                            data.tickBurning(player);
                         }
 
 
@@ -244,11 +236,11 @@ public class CommonEventHandler {
                          * CHROMIUM (enhanced)                       *
                          *********************************************/
                         if (data.isEnhanced() && data.isBurning(Metal.CHROMIUM)) {
-                            if (event.world instanceof ServerWorld) {
+                            if (level instanceof ServerLevel) {
                                 int max = 20;
                                 BlockPos negative = new BlockPos(curPlayer.position()).offset(-max, -max, -max);
                                 BlockPos positive = new BlockPos(curPlayer.position()).offset(max, max, max);
-                                event.world.getEntitiesOfClass(PlayerEntity.class, new AxisAlignedBB(negative, positive)).forEach(otherPlayer -> {
+                                level.getEntitiesOfClass(Player.class, new AABB(negative, positive)).forEach(otherPlayer -> {
                                     otherPlayer.getCapability(AllomancerCapability.PLAYER_CAP).ifPresent(otherData -> otherData.drainMetals(Metal.values()));
                                 });
                             }
@@ -259,19 +251,18 @@ public class CommonEventHandler {
                          * GOLD AND ELECTRUM (enhanced)              *
                          *********************************************/
                         if (data.isEnhanced() && data.isBurning(Metal.ELECTRUM) && data.getAmount(Metal.ELECTRUM) >= 9) {
-                            RegistryKey<World> spawnDim = data.getSpawnDim();
+                            ResourceKey<Level> spawnDim = data.getSpawnDim();
                             BlockPos spawnLoc;
 
                             if (spawnDim != null) {
                                 spawnLoc = data.getSpawnLoc();
                             } else {
-                                spawnDim = World.OVERWORLD; // no spawn --> use world spawn
-                                spawnLoc = new BlockPos(curPlayer.level.getLevelData().getXSpawn(), curPlayer.level.getLevelData().getYSpawn(),
-                                                        curPlayer.level.getLevelData().getZSpawn());
+                                spawnDim = Level.OVERWORLD; // no spawn --> use world spawn
+                                spawnLoc = new BlockPos(level.getLevelData().getXSpawn(), level.getLevelData().getYSpawn(), level.getLevelData().getZSpawn());
 
                             }
 
-                            PowerUtils.teleport(curPlayer, event.world, spawnDim, spawnLoc);
+                            PowerUtils.teleport(curPlayer, level, spawnDim, spawnLoc);
                             if (data.isBurning(Metal.DURALUMIN)) {
                                 data.drainMetals(Metal.DURALUMIN);
                             }
@@ -279,9 +270,9 @@ public class CommonEventHandler {
 
 
                         } else if (data.isEnhanced() && data.isBurning(Metal.GOLD) && data.getAmount(Metal.GOLD) >= 9) { // These should be mutually exclusive
-                            RegistryKey<World> deathDim = data.getDeathDim();
+                            ResourceKey<Level> deathDim = data.getDeathDim();
                             if (deathDim != null) {
-                                PowerUtils.teleport(curPlayer, event.world, deathDim, data.getDeathLoc());
+                                PowerUtils.teleport(curPlayer, level, deathDim, data.getDeathLoc());
                                 if (data.isBurning(Metal.DURALUMIN)) {
                                     data.drainMetals(Metal.DURALUMIN);
                                 }
@@ -294,26 +285,26 @@ public class CommonEventHandler {
                          * BENDALLOY AND CADMIUM                     *
                          *********************************************/
                         if (data.isBurning(Metal.BENDALLOY) && !data.isBurning(Metal.CADMIUM)) {
-                            curPlayer.addEffect(new EffectInstance(Effects.DIG_SPEED, 10, 3, true, false));
+                            curPlayer.addEffect(new MobEffectInstance(MobEffects.DIG_SPEED, 10, 3, true, false));
                             curPlayer.aiStep();
                             curPlayer.aiStep();
 
-                            if (event.world instanceof ServerWorld) {
+                            if (level instanceof ServerLevel serverLevel) {
                                 int max = data.isEnhanced() ? 10 : 5;
                                 BlockPos negative = new BlockPos(curPlayer.position()).offset(-max, -max, -max);
                                 BlockPos positive = new BlockPos(curPlayer.position()).offset(max, max, max);
-                                event.world.getEntitiesOfClass(LivingEntity.class, new AxisAlignedBB(negative, positive)).forEach(entity -> {
+                                serverLevel.getEntitiesOfClass(LivingEntity.class, new AABB(negative, positive)).forEach(entity -> {
                                     entity.aiStep();
                                     entity.aiStep();
                                 });
                                 BlockPos.betweenClosedStream(negative, positive).forEach(bp -> {
                                     BlockState block = event.world.getBlockState(bp);
-                                    TileEntity te = event.world.getBlockEntity(bp);
+                                    BlockEntity te = event.world.getBlockEntity(bp);
                                     for (int i = 0; i < max * 4 / (te == null ? 10 : 1); i++) {
-                                        if (te instanceof ITickableTileEntity) {
-                                            ((ITickableTileEntity) te).tick();
+                                        if (te instanceof TickingBlockEntity tbe) {
+                                            tbe.tick();
                                         } else if (block.isRandomlyTicking()) {
-                                            block.randomTick((ServerWorld) event.world, bp, event.world.random);
+                                            block.randomTick(serverLevel, bp, serverLevel.random);
                                         }
                                     }
                                 });
@@ -324,10 +315,10 @@ public class CommonEventHandler {
                             BlockPos negative = new BlockPos(curPlayer.position()).offset(-max, -max, -max);
                             BlockPos positive = new BlockPos(curPlayer.position()).offset(max, max, max);
                             int slowness_amplifier = data.isEnhanced() ? 255 : 2; // Duralumin freezes entities
-                            event.world.getEntitiesOfClass(LivingEntity.class, new AxisAlignedBB(negative, positive)).forEach(entity -> {
-                                entity.addEffect(new EffectInstance(Effects.SLOW_FALLING, 10, 0, true, false));
+                            level.getEntitiesOfClass(LivingEntity.class, new AABB(negative, positive)).forEach(entity -> {
+                                entity.addEffect(new MobEffectInstance(MobEffects.SLOW_FALLING, 10, 0, true, false));
                                 if (entity != curPlayer) {
-                                    entity.addEffect(new EffectInstance(Effects.MOVEMENT_SLOWDOWN, 10, slowness_amplifier, true, false));
+                                    entity.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SLOWDOWN, 10, slowness_amplifier, true, false));
                                 }
                             });
                         }
@@ -338,31 +329,32 @@ public class CommonEventHandler {
                          *********************************************/
                         if (data.isBurning(Metal.TIN)) {
                             // Add night vision to tin-burners
-                            curPlayer.addEffect(new EffectInstance(Effects.NIGHT_VISION, Short.MAX_VALUE, 5, true, false));
+                            curPlayer.addEffect(new MobEffectInstance(MobEffects.NIGHT_VISION, Short.MAX_VALUE, 5, true, false));
                             if (data.isEnhanced()) { // Tin and Duralumin is too much to handle
-                                curPlayer.addEffect(new EffectInstance(Effects.BLINDNESS, 100, 150, true, false));
-                                if (event.world.random.nextInt(50) == 0) {
-                                    curPlayer.addEffect(new EffectInstance(Effects.CONFUSION, 100, 0, true, false));
+                                curPlayer.addEffect(new MobEffectInstance(MobEffects.BLINDNESS, 100, 150, true, false));
+                                if (level.random.nextInt(50) == 0) {
+                                    curPlayer.addEffect(new MobEffectInstance(MobEffects.CONFUSION, 100, 0, true, false));
                                 }
                             } else { // Remove blindness from normal tin burners
-                                if (curPlayer.hasEffect(Effects.BLINDNESS)) {
-                                    curPlayer.removeEffect(Effects.BLINDNESS);
+                                if (curPlayer.hasEffect(MobEffects.BLINDNESS)) {
+                                    curPlayer.removeEffect(MobEffects.BLINDNESS);
                                 }
                             }
                         }
                         // Remove night vision from non-tin burners if duration < 10 seconds. Related to the above issue with flashing, only if the amplifier is 5
-                        if (!data.isBurning(Metal.TIN) && curPlayer.getEffect(Effects.NIGHT_VISION) != null && curPlayer.getEffect(Effects.NIGHT_VISION).getAmplifier() == 5) {
+                        if (!data.isBurning(Metal.TIN) && curPlayer.getEffect(MobEffects.NIGHT_VISION) != null &&
+                            curPlayer.getEffect(MobEffects.NIGHT_VISION).getAmplifier() == 5) {
 
-                            curPlayer.removeEffect(Effects.NIGHT_VISION);
+                            curPlayer.removeEffect(MobEffects.NIGHT_VISION);
                         }
                         if (data.isBurning(Metal.PEWTER)) {
                             //Add jump boost and speed to pewter burners
-                            curPlayer.addEffect(new EffectInstance(Effects.JUMP, 10, 1, true, false));
-                            curPlayer.addEffect(new EffectInstance(Effects.MOVEMENT_SPEED, 10, 0, true, false));
-                            curPlayer.addEffect(new EffectInstance(Effects.DIG_SPEED, 10, 1, true, false));
+                            curPlayer.addEffect(new MobEffectInstance(MobEffects.JUMP, 10, 1, true, false));
+                            curPlayer.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SPEED, 10, 0, true, false));
+                            curPlayer.addEffect(new MobEffectInstance(MobEffects.DIG_SPEED, 10, 1, true, false));
 
                             if (data.getDamageStored() > 0) {
-                                if (event.world.random.nextInt(200) == 0) {
+                                if (level.random.nextInt(200) == 0) {
                                     data.setDamageStored(data.getDamageStored() - 1);
                                 }
                             }
@@ -379,27 +371,26 @@ public class CommonEventHandler {
                          * COPPER (enhanced)                      *
                          *********************************************/
                         if (data.isEnhanced() && data.isBurning(Metal.COPPER)) {
-                            curPlayer.addEffect(new EffectInstance(Effects.INVISIBILITY, 20, 50, true, false));
+                            curPlayer.addEffect(new MobEffectInstance(MobEffects.INVISIBILITY, 20, 50, true, false));
                         }
 
-                        if (false) { //todo
-                            BlockPos negative = curPlayer.blockPosition().offset(-30, -30, -30);
-                            BlockPos positive = curPlayer.blockPosition().offset(30, 30, 30);
-                            List<MobEntity> nearby_players = curPlayer.level.getEntitiesOfClass(MobEntity.class, new AxisAlignedBB(negative, positive), Objects::nonNull);
-
-                            for (MobEntity mob : nearby_players) {
-                                Path path = mob.getNavigation().getPath();
-                                if (path != null) {
-                                    int count = path.getNodeCount();
-                                    for (int i = 0; i < count; i++) {
-                                        PathPoint point = path.getNode(i);
-                                        System.out.println(point);
-                                        // TODO this would be atium? need a packet and a renderer
-                                        event.world.addParticle(ParticleTypes.EXPLOSION, point.x + 0.5, point.y + 0.5, point.z + 0.5, 0, 0, 0);
-                                    }
-                                }
-                            }
-                        }
+                        // TODO this would be atium? need a packet and a renderer
+                        //                        if (false) {
+                        //                            BlockPos negative = curPlayer.blockPosition().offset(-30, -30, -30);
+                        //                            BlockPos positive = curPlayer.blockPosition().offset(30, 30, 30);
+                        //                            var nearby_players = curPlayer.level.getEntitiesOfClass(Mob.class, new AABB(negative, positive), Objects::nonNull);
+                        //
+                        //                            for (Mob mob : nearby_players) {
+                        //                                Path path = mob.getNavigation().getPath();
+                        //                                if (path != null) {
+                        //                                    int count = path.getNodeCount();
+                        //                                    for (int i = 0; i < count; i++) {
+                        //                                        Node point = path.getNode(i);
+                        //                                        event.world.addParticle(ParticleTypes.EXPLOSION, point.x + 0.5, point.y + 0.5, point.z + 0.5, 0, 0, 0);
+                        //                                    }
+                        //                                }
+                        //                            }
+                        //                        }
 
                     }
                 });
