@@ -1,45 +1,48 @@
 package com.legobmw99.allomancy.modules.powers;
 
 import com.legobmw99.allomancy.Allomancy;
+import com.legobmw99.allomancy.api.enums.Metal;
 import com.legobmw99.allomancy.modules.combat.entity.ProjectileNuggetEntity;
 import com.legobmw99.allomancy.modules.powers.data.AllomancerCapability;
 import com.legobmw99.allomancy.modules.powers.entity.ai.AIAttackOnCollideExtended;
 import com.legobmw99.allomancy.modules.powers.entity.ai.AIEvilAttack;
 import com.legobmw99.allomancy.network.Network;
-import com.legobmw99.allomancy.api.enums.Metal;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.entity.CreatureEntity;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.ai.attributes.Attributes;
-import net.minecraft.entity.ai.goal.*;
-import net.minecraft.entity.item.FallingBlockEntity;
-import net.minecraft.entity.item.ItemEntity;
-import net.minecraft.entity.item.ItemFrameEntity;
-import net.minecraft.entity.item.minecart.AbstractMinecartEntity;
-import net.minecraft.entity.merchant.IReputationType;
-import net.minecraft.entity.merchant.villager.AbstractVillagerEntity;
-import net.minecraft.entity.merchant.villager.VillagerEntity;
-import net.minecraft.entity.merchant.villager.WanderingTraderEntity;
-import net.minecraft.entity.monster.*;
-import net.minecraft.entity.passive.IronGolemEntity;
-import net.minecraft.entity.passive.RabbitEntity;
-import net.minecraft.entity.passive.SheepEntity;
-import net.minecraft.entity.passive.TameableEntity;
-import net.minecraft.entity.passive.horse.AbstractHorseEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.player.ServerPlayerEntity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.Hand;
-import net.minecraft.util.RegistryKey;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.MathHelper;
-import net.minecraft.util.math.vector.Vector3d;
-import net.minecraft.world.Explosion;
-import net.minecraft.world.World;
-import net.minecraft.world.server.ServerWorld;
+import net.minecraft.core.BlockPos;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.util.Mth;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.PathfinderMob;
+import net.minecraft.world.entity.TamableAnimal;
+import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.entity.ai.goal.*;
+import net.minecraft.world.entity.ai.goal.target.HurtByTargetGoal;
+import net.minecraft.world.entity.ai.goal.target.NearestAttackableTargetGoal;
+import net.minecraft.world.entity.ai.goal.target.TargetGoal;
+import net.minecraft.world.entity.ai.village.ReputationEventType;
+import net.minecraft.world.entity.animal.IronGolem;
+import net.minecraft.world.entity.animal.Rabbit;
+import net.minecraft.world.entity.animal.Sheep;
+import net.minecraft.world.entity.animal.horse.AbstractHorse;
+import net.minecraft.world.entity.decoration.ItemFrame;
+import net.minecraft.world.entity.item.FallingBlockEntity;
+import net.minecraft.world.entity.item.ItemEntity;
+import net.minecraft.world.entity.monster.*;
+import net.minecraft.world.entity.npc.AbstractVillager;
+import net.minecraft.world.entity.npc.Villager;
+import net.minecraft.world.entity.npc.WanderingTrader;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.vehicle.AbstractMinecart;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Explosion;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.common.util.ITeleporter;
 
 import java.util.function.Function;
@@ -54,7 +57,7 @@ public class PowerUtils {
 
     public static final byte PUSH = 1;
     public static final byte PULL = -1;
-    private static final Predicate<Goal> isAggroGoal = (goal) -> goal instanceof CreeperSwellGoal || goal instanceof AIAttackOnCollideExtended || goal instanceof MeleeAttackGoal ||
+    private static final Predicate<Goal> isAggroGoal = (goal) -> goal instanceof SwellGoal || goal instanceof AIAttackOnCollideExtended || goal instanceof MeleeAttackGoal ||
                                                                  goal instanceof TargetGoal || goal instanceof PanicGoal || goal.getClass().getName().contains("Fireball") ||
                                                                  goal.getClass().getName().contains("Attack") || goal.getClass().getName().contains("Anger");
 
@@ -71,7 +74,7 @@ public class PowerUtils {
      * Block state wrapper on {@link PowerUtils#isBlockMetal}
      *
      * @param state BlockState to check
-     * @return whether or not the block state is metal
+     * @return whether the block state is metal
      */
     public static boolean isBlockStateMetal(BlockState state) {
         return isBlockMetal(state.getBlock());
@@ -81,7 +84,7 @@ public class PowerUtils {
      * Determines if a block is metal or not
      *
      * @param block to be checked
-     * @return Whether or not the block is metal
+     * @return Whether the block is metal
      */
     public static boolean isBlockMetal(Block block) {
         return isOnWhitelist(block.getRegistryName().toString());
@@ -91,7 +94,7 @@ public class PowerUtils {
      * Determines if an item is metal or not
      *
      * @param item to be checked
-     * @return Whether or not the item is metal
+     * @return Whether the item is metal
      */
     public static boolean isItemMetal(ItemStack item) {
         return isOnWhitelist(item.getItem().getRegistryName().toString());
@@ -105,33 +108,32 @@ public class PowerUtils {
      * Determines if an entity is metal or not
      *
      * @param entity to be checked
-     * @return Whether or not the entity is metallic
+     * @return Whether the entity is metallic
      */
     public static boolean isEntityMetal(Entity entity) {
         if (entity == null) {
             return false;
         }
-        if (entity instanceof ItemEntity) {
-            return isItemMetal(((ItemEntity) entity).getItem());
+        if (entity instanceof ItemEntity item) {
+            return isItemMetal(item.getItem());
         }
-        if (entity instanceof ItemFrameEntity) {
-            return isItemMetal(((ItemFrameEntity) entity).getItem());
+        if (entity instanceof ItemFrame itemFrame) {
+            return isItemMetal(itemFrame.getItem());
         }
-        if (entity instanceof FallingBlockEntity) {
-            return isBlockStateMetal(((FallingBlockEntity) entity).getBlockState());
+        if (entity instanceof FallingBlockEntity fbe) {
+            return isBlockStateMetal(fbe.getBlockState());
         }
         if (entity instanceof ProjectileNuggetEntity) {
             return true;
         }
-        if (entity instanceof AbstractMinecartEntity) {
+        if (entity instanceof AbstractMinecart) {
             return true;
         }
-        if (entity instanceof LivingEntity) {
-            LivingEntity ent = (LivingEntity) entity;
-            if (ent instanceof IronGolemEntity) {
+        if (entity instanceof LivingEntity ent) {
+            if (ent instanceof IronGolem) {
                 return true;
             }
-            if (isItemMetal(ent.getItemInHand(Hand.MAIN_HAND)) || isItemMetal(ent.getItemInHand(Hand.OFF_HAND))) {
+            if (isItemMetal(ent.getItemInHand(InteractionHand.MAIN_HAND)) || isItemMetal(ent.getItemInHand(InteractionHand.OFF_HAND))) {
                 return true;
             }
             for (ItemStack itemStack : ent.getArmorSlots()) {
@@ -149,12 +151,12 @@ public class PowerUtils {
      *
      * @param player The player to wipe
      */
-    public static void wipePlayer(PlayerEntity player) {
+    public static void wipePlayer(Player player) {
         player.getCapability(AllomancerCapability.PLAYER_CAP).ifPresent(data -> {
             data.drainMetals(Metal.values());
             player.removeAllEffects();
 
-            if (player instanceof ServerPlayerEntity) {
+            if (player instanceof ServerPlayer) {
                 Network.sync(data, player);
             }
         });
@@ -173,13 +175,13 @@ public class PowerUtils {
             toMove = toMove.getVehicle();
         }
 
-        Vector3d motion = toMove.position().subtract(Vector3d.atCenterOf(block)).normalize().scale(directionScalar * 1.1);
-        Vector3d mod = clamp(cutoff(motion.add(toMove.getDeltaMovement()), 0.1), abs(motion).reverse(), abs(motion));
+        Vec3 motion = toMove.position().subtract(Vec3.atCenterOf(block)).normalize().scale(directionScalar * 1.1);
+        Vec3 mod = clamp(cutoff(motion.add(toMove.getDeltaMovement()), 0.1), abs(motion).reverse(), abs(motion));
         toMove.setDeltaMovement(mod);
         toMove.hurtMarked = true;
 
         // Only save players from fall damage
-        if (toMove instanceof ServerPlayerEntity) {
+        if (toMove instanceof ServerPlayer) {
             toMove.fallDistance = 0;
         }
     }
@@ -188,17 +190,17 @@ public class PowerUtils {
     /*
      * Three helper functions for working with Vector3ds
      */
-    private static Vector3d clamp(Vector3d value, Vector3d min, Vector3d max) {
-        return new Vector3d(MathHelper.clamp(value.x, min.x, max.x), MathHelper.clamp(value.y, min.y, max.y), MathHelper.clamp(value.z, min.z, max.z));
+    private static Vec3 clamp(Vec3 value, Vec3 min, Vec3 max) {
+        return new Vec3(Mth.clamp(value.x, min.x, max.x), Mth.clamp(value.y, min.y, max.y), Mth.clamp(value.z, min.z, max.z));
     }
 
-    private static Vector3d abs(Vector3d vec) {
-        return new Vector3d(Math.abs(vec.x), Math.abs(vec.y), Math.abs(vec.z));
+    private static Vec3 abs(Vec3 vec) {
+        return new Vec3(Math.abs(vec.x), Math.abs(vec.y), Math.abs(vec.z));
     }
 
-    private static Vector3d cutoff(Vector3d value, double e) {
-        Vector3d mag = abs(value);
-        return new Vector3d(mag.x < e ? 0 : value.x, mag.y < e ? 0 : value.y, mag.z < e ? 0 : value.z);
+    private static Vec3 cutoff(Vec3 value, double e) {
+        Vec3 mag = abs(value);
+        return new Vec3(mag.x < e ? 0 : value.x, mag.y < e ? 0 : value.y, mag.z < e ? 0 : value.z);
     }
 
 
@@ -210,7 +212,7 @@ public class PowerUtils {
      * @param dimension Dimension to call {@link Entity#changeDimension} on
      * @param pos       BlockPos to move the player to using {@link Entity#teleportToWithTicket}
      */
-    public static void teleport(PlayerEntity player, World world, RegistryKey<World> dimension, BlockPos pos) {
+    public static void teleport(Player player, Level world, ResourceKey<Level> dimension, BlockPos pos) {
         if (!world.isClientSide) {
             if (player != null) {
                 if (player.isPassenger()) {
@@ -219,9 +221,9 @@ public class PowerUtils {
 
                 if (player.level.dimension() != dimension) {
                     //change dimension
-                    player = (PlayerEntity) player.changeDimension(world.getServer().getLevel(dimension), new ITeleporter() {
+                    player = (Player) player.changeDimension(world.getServer().getLevel(dimension), new ITeleporter() {
                         @Override
-                        public Entity placeEntity(Entity entity, ServerWorld currentWorld, ServerWorld destWorld, float yaw, Function<Boolean, Entity> repositionEntity) {
+                        public Entity placeEntity(Entity entity, ServerLevel currentWorld, ServerLevel destWorld, float yaw, Function<Boolean, Entity> repositionEntity) {
                             Entity repositionedEntity = repositionEntity.apply(false);
                             repositionedEntity.teleportTo(pos.getX(), pos.getY(), pos.getZ());
                             return repositionedEntity;
@@ -235,7 +237,7 @@ public class PowerUtils {
         }
     }
 
-    public static void riotEntity(CreatureEntity target, PlayerEntity allomancer, boolean enhanced) {
+    public static void riotEntity(PathfinderMob target, Player allomancer, boolean enhanced) {
         try {
             if (!enhanced) {
                 //Enable Targeting goals
@@ -245,49 +247,49 @@ public class PowerUtils {
                 target.setLastHurtByMob(allomancer);
                 // TODO: try to use PrioritizedGoal::startExecuting for already hostiles
                 target.targetSelector.addGoal(1, new AIAttackOnCollideExtended(target, 1d, false));
-                target.targetSelector.addGoal(2, new NearestAttackableTargetGoal<>(target, PlayerEntity.class, false));
+                target.targetSelector.addGoal(2, new NearestAttackableTargetGoal<>(target, Player.class, false));
                 target.targetSelector.addGoal(1, new NearestAttackableTargetGoal<>(target, target.getClass(), false));
-                target.goalSelector.addGoal(4, new LookRandomlyGoal(target));
+                target.goalSelector.addGoal(4, new RandomLookAroundGoal(target));
                 target.targetSelector.addGoal(2, new HurtByTargetGoal(target).setAlertOthers());
-                if (target.getAttribute(Attributes.ATTACK_DAMAGE) != null && !(target instanceof GuardianEntity)) {
+                if (target.getAttribute(Attributes.ATTACK_DAMAGE) != null && !(target instanceof Guardian)) {
                     target.goalSelector.addGoal(3, new MeleeAttackGoal(target, 1.2D, true));
                 }
 
                 target.setAggressive(true);
 
-                if (target instanceof CreeperEntity) {
-                    target.goalSelector.addGoal(1, new CreeperSwellGoal((CreeperEntity) target));
+                if (target instanceof Creeper creeper) {
+                    target.goalSelector.addGoal(1, new SwellGoal(creeper));
                 }
-                if (target instanceof RabbitEntity) {
-                    target.goalSelector.addGoal(1, new AIEvilAttack((RabbitEntity) target));
+                if (target instanceof Rabbit rabbit) {
+                    target.goalSelector.addGoal(1, new AIEvilAttack(rabbit));
                 }
-                if (target instanceof AbstractSkeletonEntity) {
-                    target.goalSelector.addGoal(1, new RangedBowAttackGoal<>((AbstractSkeletonEntity) target, 1.0D, 20, 15.0F));
+                if (target instanceof AbstractSkeleton skeleton) {
+                    target.goalSelector.addGoal(1, new RangedBowAttackGoal<>(skeleton, 1.0D, 20, 15.0F));
                 }
-                if (target instanceof IllusionerEntity) {
-                    target.goalSelector.addGoal(1, new RangedBowAttackGoal<>((IllusionerEntity) target, 0.5D, 20, 15.0F));
+                if (target instanceof Illusioner illusioner) {
+                    target.goalSelector.addGoal(1, new RangedBowAttackGoal<>(illusioner, 0.5D, 20, 15.0F));
                 }
-                if (target instanceof PillagerEntity) {
-                    target.goalSelector.addGoal(2, new RangedCrossbowAttackGoal<>((PillagerEntity) target, 1.0D, 8.0F));
+                if (target instanceof Pillager pillager) {
+                    target.goalSelector.addGoal(2, new RangedCrossbowAttackGoal<>(pillager, 1.0D, 8.0F));
                 }
             } else {
-                target.level.explode(target, target.position().x(), target.position().y(), target.position().z(), 1.2F, false, Explosion.Mode.BREAK);
-                target.remove();
+                target.level.explode(target, target.position().x(), target.position().y(), target.position().z(), 1.2F, false, Explosion.BlockInteraction.BREAK);
+                target.kill();
             }
         } catch (Exception e) {
             Allomancy.LOGGER.error("Failed to riot entity " + target + "! Please report this error!", e);
         }
     }
 
-    public static void sootheEntity(CreatureEntity target, PlayerEntity allomancer, boolean enhanced) {
+    public static void sootheEntity(PathfinderMob target, Player allomancer, boolean enhanced) {
         try {
             if (!enhanced) {
                 if (target.isNoAi()) {
                     target.setNoAi(false);
                 }
                 // Reset all current aggro goals
-                target.goalSelector.getRunningGoals().filter(isAggroGoal).forEach(PrioritizedGoal::stop);
-                target.targetSelector.getRunningGoals().filter(isAggroGoal).forEach(PrioritizedGoal::stop);
+                target.goalSelector.getRunningGoals().filter(isAggroGoal).forEach(WrappedGoal::stop);
+                target.targetSelector.getRunningGoals().filter(isAggroGoal).forEach(WrappedGoal::stop);
                 target.goalSelector.tick();
                 target.targetSelector.tick();
                 target.setTarget(null);
@@ -296,28 +298,28 @@ public class PowerUtils {
                 target.targetSelector.disableControlFlag(Goal.Flag.TARGET);
                 target.setAggressive(false);
                 //Add new goals
-                target.goalSelector.addGoal(7, new LookAtGoal(target, PlayerEntity.class, 6.0F));
+                target.goalSelector.addGoal(7, new LookAtPlayerGoal(target, Player.class, 6.0F));
 
-                if (target instanceof TameableEntity) {
+                if (target instanceof TamableAnimal animal) {
                     if (Math.random() < 0.3) {
-                        ((TameableEntity) target).tame(allomancer);
+                        animal.tame(allomancer);
                     }
                 }
-                if (target instanceof AbstractHorseEntity) {
+                if (target instanceof AbstractHorse horse) {
                     if (Math.random() < 0.3) {
-                        ((AbstractHorseEntity) target).tameWithName(allomancer);
+                        horse.tameWithName(allomancer);
                     }
                 }
-                if (target instanceof SheepEntity) {
-                    target.goalSelector.addGoal(1, new EatGrassGoal(target));
+                if (target instanceof Sheep) {
+                    target.goalSelector.addGoal(1, new EatBlockGoal(target));
                 }
-                if (target instanceof VillagerEntity) {
-                    ((VillagerEntity) target).onReputationEventFrom(IReputationType.TRADE, allomancer);
+                if (target instanceof Villager villager) {
+                    villager.onReputationEventFrom(ReputationEventType.TRADE, allomancer);
                 }
-                if (target instanceof WanderingTraderEntity) {
-                    target.goalSelector.addGoal(1, new TradeWithPlayerGoal((AbstractVillagerEntity) target));
+                if (target instanceof WanderingTrader) {
+                    target.goalSelector.addGoal(1, new TradeWithPlayerGoal((AbstractVillager) target));
                 }
-            } else { // Completely remove all ai if enhanced
+            } else { // Completely remove all AI if enhanced
                 target.setNoAi(true);
             }
 
