@@ -1,10 +1,10 @@
 package com.legobmw99.allomancy.api.block;
 
 import com.legobmw99.allomancy.modules.extras.block.IronLeverBlock;
-import net.minecraft.block.BlockState;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.state.BlockState;
 
 /**
  * This interface can be used to signify if a block should react to being pushed
@@ -18,8 +18,8 @@ public interface IAllomanticallyUsableBlock {
     /**
      * Called when the block is steelpushed or ironpulled
      *
-     * @param isPush whether or not the activation is Steel
-     * @return whether or not the block was activated
+     * @param isPush whether the activation is Steel
+     * @return whether the block was activated
      */
-    boolean useAllomantically(BlockState state, World world, BlockPos pos, PlayerEntity playerIn, boolean isPush);
+    boolean useAllomantically(BlockState state, Level world, BlockPos pos, Player playerIn, boolean isPush);
 }
