@@ -8,13 +8,16 @@ import com.legobmw99.allomancy.modules.consumables.ConsumeSetup;
 import net.minecraft.advancements.Advancement;
 import net.minecraft.advancements.AdvancementRewards;
 import net.minecraft.advancements.FrameType;
-import net.minecraft.advancements.criterion.ConsumeItemTrigger;
-import net.minecraft.data.AdvancementProvider;
+import net.minecraft.advancements.critereon.ConsumeItemTrigger;
+import net.minecraft.advancements.critereon.ImpossibleTrigger;
+import net.minecraft.advancements.critereon.InventoryChangeTrigger;
 import net.minecraft.data.DataGenerator;
-import net.minecraft.data.DirectoryCache;
-import net.minecraft.data.IDataProvider;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.text.TranslationTextComponent;
+import net.minecraft.data.DataProvider;
+import net.minecraft.data.HashCache;
+import net.minecraft.data.advancements.AdvancementProvider;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraftforge.common.data.ExistingFileHelper;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -25,31 +28,50 @@ public class Advancements extends AdvancementProvider {
 
     private final DataGenerator gen;
 
-    public Advancements(DataGenerator generatorIn) {
-        super(generatorIn);
+    public Advancements(DataGenerator generatorIn, ExistingFileHelper existingFileHelper) {
+        super(generatorIn, existingFileHelper);
         this.gen = generatorIn;
     }
 
     private static void registerAdvancements(Consumer<Advancement> consumer) {
+
+
         Advancement.Builder
                 .advancement()
-                .parent(Advancement.Builder.advancement().build(new ResourceLocation("adventure/root"))) // hacky
-                .display(CombatSetup.MISTCLOAK.get(), new TranslationTextComponent("advancements.become_mistborn.title"),
-                         new TranslationTextComponent("advancements.become_mistborn.desc"), null, FrameType.CHALLENGE, true, true, true)
-                .addCriterion("lerasium_nugget", ConsumeItemTrigger.Instance.usedItem(ConsumeSetup.LERASIUM_NUGGET.get()))
+                .parent(Advancement.Builder.advancement().build(new ResourceLocation("adventure/root")))
+                .display(ConsumeSetup.ALLOMANTIC_GRINDER.get(), new TranslatableComponent("advancements.local_metallurgist.title"),
+                         new TranslatableComponent("advancements.local_metallurgist.desc"), null, FrameType.TASK, true, true, false)
+                .addCriterion("grinder", InventoryChangeTrigger.TriggerInstance.hasItems(ConsumeSetup.ALLOMANTIC_GRINDER.get()))
+                .save(consumer, "allomancy:main/metallurgist");
+
+        Advancement.Builder
+                .advancement()
+                .parent(Advancement.Builder.advancement().build(new ResourceLocation(Allomancy.MODID, "main/metallurgist")))
+                .display(ConsumeSetup.LERASIUM_NUGGET.get(), new TranslatableComponent("advancements.dna_entangled.title"),
+                         new TranslatableComponent("advancements.dna_entangled.desc"), null, FrameType.TASK, true, false, true)
+                .addCriterion("impossible", new ImpossibleTrigger.TriggerInstance())
+                .save(consumer, "allomancy:main/dna_entangled");
+
+        Advancement.Builder
+                .advancement()
+                .parent(Advancement.Builder.advancement().build(new ResourceLocation(Allomancy.MODID, "main/metallurgist")))
+                .display(CombatSetup.MISTCLOAK.get(), new TranslatableComponent("advancements.become_mistborn.title"),
+                         new TranslatableComponent("advancements.become_mistborn.desc"), null, FrameType.CHALLENGE, true, true, true)
+                .addCriterion("lerasium_nugget", ConsumeItemTrigger.TriggerInstance.usedItem(ConsumeSetup.LERASIUM_NUGGET.get()))
                 .rewards(AdvancementRewards.Builder.experience(100))
                 .save(consumer, "allomancy:main/become_mistborn");
+
 
     }
 
     @Override
-    public void run(DirectoryCache cache) {
+    public void run(HashCache cache) {
         Path outputFolder = this.gen.getOutputFolder();
         Consumer<Advancement> consumer = (advancement) -> {
 
             Path path = outputFolder.resolve("data/" + advancement.getId().getNamespace() + "/advancements/" + advancement.getId().getPath() + ".json");
             try {
-                IDataProvider.save(GSON, cache, advancement.deconstruct().serializeToJson(), path);
+                DataProvider.save(GSON, cache, advancement.deconstruct().serializeToJson(), path);
                 Allomancy.LOGGER.debug("Creating advancement " + advancement.getId());
             } catch (IOException ioexception) {
                 Allomancy.LOGGER.error("Couldn't save advancement {}", path, ioexception);
