@@ -1,24 +1,22 @@
 package com.legobmw99.allomancy.modules.combat;
 
 import com.legobmw99.allomancy.Allomancy;
-import com.legobmw99.allomancy.modules.combat.client.CombatClientSetup;
 import com.legobmw99.allomancy.modules.combat.entity.ProjectileNuggetEntity;
 import com.legobmw99.allomancy.modules.combat.item.CoinBagItem;
 import com.legobmw99.allomancy.modules.combat.item.KolossBladeItem;
 import com.legobmw99.allomancy.modules.combat.item.MistcloakItem;
 import com.legobmw99.allomancy.modules.combat.item.ObsidianDaggerItem;
-import net.minecraft.entity.EntityClassification;
-import net.minecraft.entity.EntityType;
-import net.minecraft.inventory.EquipmentSlotType;
-import net.minecraft.item.IArmorMaterial;
-import net.minecraft.item.Item;
-import net.minecraft.item.Items;
-import net.minecraft.item.crafting.Ingredient;
-import net.minecraft.util.SoundEvent;
-import net.minecraft.util.SoundEvents;
-import net.minecraftforge.fml.RegistryObject;
-import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
+import net.minecraft.sounds.SoundEvent;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.MobCategory;
+import net.minecraft.world.item.ArmorMaterial;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.fmllegacy.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 
@@ -28,25 +26,15 @@ public class CombatSetup {
     public static final RegistryObject<CoinBagItem> COIN_BAG = ITEMS.register("coin_bag", CoinBagItem::new);
     public static final RegistryObject<ObsidianDaggerItem> OBSIDIAN_DAGGER = ITEMS.register("obsidian_dagger", ObsidianDaggerItem::new);
     public static final RegistryObject<KolossBladeItem> KOLOSS_BLADE = ITEMS.register("koloss_blade", KolossBladeItem::new);
-    public static final RegistryObject<EntityType<ProjectileNuggetEntity>> NUGGET_PROJECTILE = ENTITIES.register("nugget_projectile",
-                                                                                                                 () -> EntityType.Builder.<ProjectileNuggetEntity>of(
-                                                                                                                         ProjectileNuggetEntity::new, EntityClassification.MISC)
-                                                                                                                         .setShouldReceiveVelocityUpdates(true)
-                                                                                                                         .setUpdateInterval(20)
-                                                                                                                         .setCustomClientFactory(
-                                                                                                                                 (spawnEntity, world) -> new ProjectileNuggetEntity(
-                                                                                                                                         world, spawnEntity.getEntity()))
-                                                                                                                         .sized(0.25F, 0.25F)
-                                                                                                                         .build("nugget_projectile"));
-    public static final IArmorMaterial WoolArmor = new IArmorMaterial() {
+    public static final ArmorMaterial WoolArmor = new ArmorMaterial() {
         @Override
-        public int getDurabilityForSlot(EquipmentSlotType slotIn) {
+        public int getDurabilityForSlot(EquipmentSlot slotIn) {
             return 50;
         }
 
         @Override
-        public int getDefenseForSlot(EquipmentSlotType slotIn) {
-            return slotIn == EquipmentSlotType.CHEST ? 4 : 0;
+        public int getDefenseForSlot(EquipmentSlot slotIn) {
+            return slotIn == EquipmentSlot.CHEST ? 4 : 0;
         }
 
         @Override
@@ -81,13 +69,18 @@ public class CombatSetup {
 
     };
     public static final RegistryObject<MistcloakItem> MISTCLOAK = ITEMS.register("mistcloak", MistcloakItem::new);
+    public static final RegistryObject<EntityType<ProjectileNuggetEntity>> NUGGET_PROJECTILE = ENTITIES.register("nugget_projectile", () -> EntityType.Builder
+            .<ProjectileNuggetEntity>of(ProjectileNuggetEntity::new, MobCategory.MISC)
+            .setShouldReceiveVelocityUpdates(true)
+            .setUpdateInterval(20)
+            .setCustomClientFactory((spawnEntity, world) -> new ProjectileNuggetEntity(world, spawnEntity.getEntity()))
+            .sized(0.25F, 0.25F)
+            .build("nugget_projectile"));
 
     public static void register() {
         ENTITIES.register(FMLJavaModLoadingContext.get().getModEventBus());
         ITEMS.register(FMLJavaModLoadingContext.get().getModEventBus());
     }
 
-    public static void clientInit(final FMLClientSetupEvent e) {
-        CombatClientSetup.registerEntityRenders();
-    }
+
 }
