@@ -3,29 +3,29 @@ package com.legobmw99.allomancy.modules.consumables.item;
 import com.legobmw99.allomancy.Allomancy;
 import com.legobmw99.allomancy.api.data.IAllomancerData;
 import com.legobmw99.allomancy.modules.powers.data.AllomancerCapability;
-import net.minecraft.client.util.ITooltipFlag;
-import net.minecraft.entity.EntityType;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.effect.LightningBoltEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.Food;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Rarity;
-import net.minecraft.potion.EffectInstance;
-import net.minecraft.potion.Effects;
-import net.minecraft.util.ActionResult;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.Hand;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TextFormatting;
-import net.minecraft.world.World;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LightningBolt;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.food.FoodProperties;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Rarity;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.level.Level;
 
 import javax.annotation.Nullable;
 import java.util.List;
 
 public class LerasiumItem extends Item {
-    private static final Food lerasium = new Food.Builder().fast().alwaysEat().saturationMod(0).nutrition(0).build();
+    private static final FoodProperties lerasium = new FoodProperties.Builder().fast().alwaysEat().saturationMod(0).nutrition(0).build();
 
     public LerasiumItem() {
         super(Allomancy.createStandardItemProperties().rarity(Rarity.EPIC).stacksTo(1).food(lerasium));
@@ -33,36 +33,36 @@ public class LerasiumItem extends Item {
 
 
     @Override
-    public ActionResult<ItemStack> use(World world, PlayerEntity player, Hand hand) {
+    public InteractionResultHolder<ItemStack> use(Level world, Player player, InteractionHand hand) {
         ItemStack itemStackIn = player.getItemInHand(hand);
         if (player.getCapability(AllomancerCapability.PLAYER_CAP).filter(data -> !data.isMistborn()).isPresent()) {
             player.startUsingItem(hand);
-            return new ActionResult<>(ActionResultType.SUCCESS, itemStackIn);
+            return new InteractionResultHolder<>(InteractionResult.SUCCESS, itemStackIn);
         }
 
-        return new ActionResult<>(ActionResultType.FAIL, itemStackIn);
+        return new InteractionResultHolder<>(InteractionResult.FAIL, itemStackIn);
 
     }
 
     @Override
-    public ItemStack finishUsingItem(ItemStack stack, World world, LivingEntity livingEntity) {
+    public ItemStack finishUsingItem(ItemStack stack, Level world, LivingEntity livingEntity) {
 
         livingEntity.getCapability(AllomancerCapability.PLAYER_CAP).ifPresent(IAllomancerData::setMistborn);
-        //Fancy shmancy effects
-        LightningBoltEntity lightning = new LightningBoltEntity(EntityType.LIGHTNING_BOLT, world);
+        //Fancy-shmancy effects
+        LightningBolt lightning = new LightningBolt(EntityType.LIGHTNING_BOLT, world);
         lightning.setVisualOnly(true);
         lightning.moveTo(livingEntity.position().add(0, 3, 0));
         world.addFreshEntity(lightning);
-        livingEntity.addEffect(new EffectInstance(Effects.FIRE_RESISTANCE, 20, 0, true, false));
+        livingEntity.addEffect(new MobEffectInstance(MobEffects.FIRE_RESISTANCE, 20, 0, true, false));
 
         return super.finishUsingItem(stack, world, livingEntity);
     }
 
 
     @Override
-    public void appendHoverText(ItemStack stack, @Nullable World worldIn, List<ITextComponent> tooltip, ITooltipFlag flagIn) {
+    public void appendHoverText(ItemStack stack, @Nullable Level worldIn, List<Component> tooltip, TooltipFlag flagIn) {
         super.appendHoverText(stack, worldIn, tooltip, flagIn);
-        ITextComponent lore = Allomancy.addColorToText("item.allomancy.lerasium_nugget.lore", TextFormatting.LIGHT_PURPLE);
+        Component lore = Allomancy.addColorToText("item.allomancy.lerasium_nugget.lore", ChatFormatting.LIGHT_PURPLE);
         tooltip.add(lore);
 
     }
