@@ -1,16 +1,20 @@
 package com.legobmw99.allomancy.modules.materials;
 
 import com.legobmw99.allomancy.Allomancy;
+import com.legobmw99.allomancy.api.enums.Metal;
 import com.legobmw99.allomancy.modules.materials.world.LootTableInjector;
 import com.legobmw99.allomancy.modules.materials.world.OreGenerator;
-import com.legobmw99.allomancy.api.enums.Metal;
-import net.minecraft.block.Block;
-import net.minecraft.item.BlockItem;
-import net.minecraft.item.Item;
+import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.OreBlock;
+import net.minecraft.world.level.block.SoundType;
+import net.minecraft.world.level.material.Material;
+import net.minecraft.world.level.material.MaterialColor;
 import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.fml.RegistryObject;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.fmllegacy.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 
@@ -18,6 +22,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class MaterialsSetup {
+
+    public static final String[] ORE_METALS = {"aluminum", "cadmium", "chromium", "lead", "silver", "tin", "zinc"};
+
     public static final DeferredRegister<Block> BLOCKS = DeferredRegister.create(ForgeRegistries.BLOCKS, Allomancy.MODID);
     public static final DeferredRegister<Item> ITEMS = DeferredRegister.create(ForgeRegistries.ITEMS, Allomancy.MODID);
 
@@ -27,22 +34,14 @@ public class MaterialsSetup {
     public static final List<RegistryObject<Block>> STORAGE_BLOCKS = new ArrayList<>();
     public static final List<RegistryObject<Item>> STORAGE_BLOCK_ITEMS = new ArrayList<>();
 
-    public static final RegistryObject<Block> ALUMINUM_ORE = BLOCKS.register("aluminum_ore", Allomancy::createStandardBlock);
-    public static final RegistryObject<Item> ALUMINUM_ORE_ITEM = ITEMS.register("aluminum_ore", () -> new BlockItem(ALUMINUM_ORE.get(), Allomancy.createStandardItemProperties()));
-    public static final RegistryObject<Block> CADMIUM_ORE = BLOCKS.register("cadmium_ore", Allomancy::createStandardBlock);
-    public static final RegistryObject<Item> CADMIUM_ORE_ITEM = ITEMS.register("cadmium_ore", () -> new BlockItem(CADMIUM_ORE.get(), Allomancy.createStandardItemProperties()));
-    public static final RegistryObject<Block> CHROMIUM_ORE = BLOCKS.register("chromium_ore", Allomancy::createStandardBlock);
-    public static final RegistryObject<Item> CHROMIUM_ORE_ITEM = ITEMS.register("chromium_ore", () -> new BlockItem(CHROMIUM_ORE.get(), Allomancy.createStandardItemProperties()));
-    public static final RegistryObject<Block> COPPER_ORE = BLOCKS.register("copper_ore", Allomancy::createStandardBlock);
-    public static final RegistryObject<Item> COPPER_ORE_ITEM = ITEMS.register("copper_ore", () -> new BlockItem(COPPER_ORE.get(), Allomancy.createStandardItemProperties()));
-    public static final RegistryObject<Block> LEAD_ORE = BLOCKS.register("lead_ore", Allomancy::createStandardBlock);
-    public static final RegistryObject<Item> LEAD_ORE_ITEM = ITEMS.register("lead_ore", () -> new BlockItem(LEAD_ORE.get(), Allomancy.createStandardItemProperties()));
-    public static final RegistryObject<Block> SILVER_ORE = BLOCKS.register("silver_ore", Allomancy::createStandardBlock);
-    public static final RegistryObject<Item> SILVER_ORE_ITEM = ITEMS.register("silver_ore", () -> new BlockItem(SILVER_ORE.get(), Allomancy.createStandardItemProperties()));
-    public static final RegistryObject<Block> TIN_ORE = BLOCKS.register("tin_ore", Allomancy::createStandardBlock);
-    public static final RegistryObject<Item> TIN_ORE_ITEM = ITEMS.register("tin_ore", () -> new BlockItem(TIN_ORE.get(), Allomancy.createStandardItemProperties()));
-    public static final RegistryObject<Block> ZINC_ORE = BLOCKS.register("zinc_ore", Allomancy::createStandardBlock);
-    public static final RegistryObject<Item> ZINC_ORE_ITEM = ITEMS.register("zinc_ore", () -> new BlockItem(ZINC_ORE.get(), Allomancy.createStandardItemProperties()));
+
+    public static final List<RegistryObject<OreBlock>> ORE_BLOCKS = new ArrayList<>();
+    public static final List<RegistryObject<Item>> ORE_BLOCKS_ITEMS = new ArrayList<>();
+    public static final List<RegistryObject<OreBlock>> DEEPSLATE_ORE_BLOCKS = new ArrayList<>();
+    public static final List<RegistryObject<Item>> DEEPSLATE_ORE_BLOCKS_ITEMS = new ArrayList<>();
+    public static final List<RegistryObject<Block>> RAW_ORE_BLOCKS = new ArrayList<>();
+    public static final List<RegistryObject<Item>> RAW_ORE_BLOCKS_ITEMS = new ArrayList<>();
+    public static final List<RegistryObject<Item>> RAW_ORE_ITEMS = new ArrayList<>();
 
     public static int METAL_ITEM_LEN = Metal.values().length;
     public static final int LEAD = METAL_ITEM_LEN++;
@@ -61,22 +60,37 @@ public class MaterialsSetup {
             } else {
                 NUGGETS.add(ITEMS.register(name + "_nugget", Allomancy::createStandardItem));
                 INGOTS.add(ITEMS.register(name + "_ingot", Allomancy::createStandardItem));
-                STORAGE_BLOCKS.add(BLOCKS.register(name + "_block", Allomancy::createStandardBlock));
+                STORAGE_BLOCKS.add(BLOCKS.register(name + "_block", MaterialsSetup::createStandardBlock));
                 STORAGE_BLOCK_ITEMS.add(ITEMS.register(name + "_block", () -> new BlockItem(STORAGE_BLOCKS.get(mt.getIndex()).get(), Allomancy.createStandardItemProperties())));
             }
         }
         FLAKES.add(MaterialsSetup.ITEMS.register("lead_flakes", Allomancy::createStandardItem));
         NUGGETS.add(ITEMS.register("lead_nugget", Allomancy::createStandardItem));
         INGOTS.add(ITEMS.register("lead_ingot", Allomancy::createStandardItem));
-        STORAGE_BLOCKS.add(BLOCKS.register("lead_block", Allomancy::createStandardBlock));
+        STORAGE_BLOCKS.add(BLOCKS.register("lead_block", MaterialsSetup::createStandardBlock));
         STORAGE_BLOCK_ITEMS.add(ITEMS.register("lead_block", () -> new BlockItem(STORAGE_BLOCKS.get(LEAD).get(), Allomancy.createStandardItemProperties())));
 
         FLAKES.add(MaterialsSetup.ITEMS.register("silver_flakes", Allomancy::createStandardItem));
         NUGGETS.add(ITEMS.register("silver_nugget", Allomancy::createStandardItem));
         INGOTS.add(ITEMS.register("silver_ingot", Allomancy::createStandardItem));
-        STORAGE_BLOCKS.add(BLOCKS.register("silver_block", Allomancy::createStandardBlock));
+        STORAGE_BLOCKS.add(BLOCKS.register("silver_block", MaterialsSetup::createStandardBlock));
         STORAGE_BLOCK_ITEMS.add(ITEMS.register("silver_block", () -> new BlockItem(STORAGE_BLOCKS.get(SILVER).get(), Allomancy.createStandardItemProperties())));
 
+        for (String ore : ORE_METALS) {
+            var ore_block = BLOCKS.register(ore + "_ore", MaterialsSetup::createStandardOre);
+            ORE_BLOCKS.add(ore_block);
+            ORE_BLOCKS_ITEMS.add(ITEMS.register(ore + "_ore", () -> new BlockItem(ore_block.get(), Allomancy.createStandardItemProperties())));
+
+            var ds_ore_block = BLOCKS.register("deepslate_" + ore + "_ore", MaterialsSetup::createDeepslateBlock);
+            DEEPSLATE_ORE_BLOCKS.add(ds_ore_block);
+            DEEPSLATE_ORE_BLOCKS_ITEMS.add(ITEMS.register("deepslate_" + ore + "_ore", () -> new BlockItem(ds_ore_block.get(), Allomancy.createStandardItemProperties())));
+
+            var raw_ore_block = BLOCKS.register("raw_" + ore + "_block", MaterialsSetup::createStandardBlock);
+            RAW_ORE_BLOCKS.add(raw_ore_block);
+            RAW_ORE_BLOCKS_ITEMS.add(ITEMS.register("raw_" + ore + "_block", () -> new BlockItem(raw_ore_block.get(), Allomancy.createStandardItemProperties())));
+
+            RAW_ORE_ITEMS.add(ITEMS.register("raw_" + ore, Allomancy::createStandardItem));
+        }
     }
 
 
@@ -90,6 +104,16 @@ public class MaterialsSetup {
         OreGenerator.registerFeatures();
     }
 
+    public static Block createStandardBlock() {
+        return new Block(Block.Properties.of(Material.STONE).strength(2.1F).requiresCorrectToolForDrops());
+    }
 
+    public static OreBlock createStandardOre() {
+        return new OreBlock(Block.Properties.of(Material.STONE).requiresCorrectToolForDrops().strength(3.0F, 3.0F));
+    }
+
+    public static OreBlock createDeepslateBlock() {
+        return new OreBlock(Block.Properties.of(Material.STONE).requiresCorrectToolForDrops().color(MaterialColor.DEEPSLATE).strength(4.5F, 3.0F).sound(SoundType.DEEPSLATE));
+    }
 }
 
