@@ -1,21 +1,21 @@
 package com.legobmw99.allomancy.modules.materials.world;
 
+import com.google.common.collect.ImmutableList;
 import com.legobmw99.allomancy.Allomancy;
 import com.legobmw99.allomancy.modules.materials.MaterialsSetup;
-import net.minecraft.block.Block;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.registry.Registry;
-import net.minecraft.util.registry.WorldGenRegistries;
-import net.minecraft.world.gen.GenerationStage;
-import net.minecraft.world.gen.feature.ConfiguredFeature;
-import net.minecraft.world.gen.feature.Feature;
-import net.minecraft.world.gen.feature.OreFeatureConfig;
-import net.minecraft.world.gen.placement.Placement;
-import net.minecraft.world.gen.placement.TopSolidRangeConfig;
+import net.minecraft.core.Registry;
+import net.minecraft.data.BuiltinRegistries;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.OreBlock;
+import net.minecraft.world.level.levelgen.GenerationStep;
+import net.minecraft.world.level.levelgen.VerticalAnchor;
+import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
+import net.minecraft.world.level.levelgen.feature.Feature;
+import net.minecraft.world.level.levelgen.feature.configurations.OreConfiguration;
 import net.minecraftforge.common.ForgeConfigSpec;
-import net.minecraftforge.common.world.BiomeGenerationSettingsBuilder;
 import net.minecraftforge.event.world.BiomeLoadingEvent;
-import net.minecraftforge.fml.RegistryObject;
+import net.minecraftforge.fmllegacy.RegistryObject;
 
 import java.util.ArrayList;
 
@@ -26,20 +26,24 @@ public class OreGenerator {
     private static final ArrayList<OreData> ores = new ArrayList<>();
 
     static {
-        ores.add(new OreData(aluminum_max_y, aluminum_min_y, aluminum_size, aluminum_density, MaterialsSetup.ALUMINUM_ORE, generate_aluminum));
-        ores.add(new OreData(cadmium_max_y, cadmium_min_y, cadmium_size, cadmium_density, MaterialsSetup.CADMIUM_ORE, generate_cadmium));
-        ores.add(new OreData(chromium_max_y, chromium_min_y, chromium_size, cadmium_density, MaterialsSetup.CHROMIUM_ORE, generate_chromium));
-        ores.add(new OreData(copper_max_y, copper_min_y, copper_size, copper_density, MaterialsSetup.COPPER_ORE, generate_copper));
-        ores.add(new OreData(lead_max_y, lead_min_y, lead_size, lead_density, MaterialsSetup.LEAD_ORE, generate_lead));
-        ores.add(new OreData(silver_max_y, silver_min_y, silver_size, silver_density, MaterialsSetup.SILVER_ORE, generate_silver));
-        ores.add(new OreData(tin_max_y, tin_min_y, tin_size, tin_density, MaterialsSetup.TIN_ORE, generate_tin));
-        ores.add(new OreData(zinc_max_y, zinc_min_y, zinc_size, zinc_density, MaterialsSetup.ZINC_ORE, generate_zinc));
+        int i = 0;
+        ores.add(new OreData(aluminum_max_y, aluminum_min_y, aluminum_size, aluminum_density, MaterialsSetup.ORE_BLOCKS.get(i), MaterialsSetup.DEEPSLATE_ORE_BLOCKS.get(i++),
+                             generate_aluminum));
+        ores.add(new OreData(cadmium_max_y, cadmium_min_y, cadmium_size, cadmium_density, MaterialsSetup.ORE_BLOCKS.get(i), MaterialsSetup.DEEPSLATE_ORE_BLOCKS.get(i++),
+                             generate_cadmium));
+        ores.add(new OreData(chromium_max_y, chromium_min_y, chromium_size, cadmium_density, MaterialsSetup.ORE_BLOCKS.get(i), MaterialsSetup.DEEPSLATE_ORE_BLOCKS.get(i++),
+                             generate_chromium));
+        ores.add(new OreData(lead_max_y, lead_min_y, lead_size, lead_density, MaterialsSetup.ORE_BLOCKS.get(i), MaterialsSetup.DEEPSLATE_ORE_BLOCKS.get(i++), generate_lead));
+        ores.add(new OreData(silver_max_y, silver_min_y, silver_size, silver_density, MaterialsSetup.ORE_BLOCKS.get(i), MaterialsSetup.DEEPSLATE_ORE_BLOCKS.get(i++),
+                             generate_silver));
+        ores.add(new OreData(tin_max_y, tin_min_y, tin_size, tin_density, MaterialsSetup.ORE_BLOCKS.get(i), MaterialsSetup.DEEPSLATE_ORE_BLOCKS.get(i++), generate_tin));
+        ores.add(new OreData(zinc_max_y, zinc_min_y, zinc_size, zinc_density, MaterialsSetup.ORE_BLOCKS.get(i), MaterialsSetup.DEEPSLATE_ORE_BLOCKS.get(i), generate_zinc));
     }
 
     public static void registerFeatures() {
-        Registry<ConfiguredFeature<?, ?>> registry = WorldGenRegistries.CONFIGURED_FEATURE;
+        var registry = BuiltinRegistries.CONFIGURED_FEATURE;
         for (OreData ore : ores) {
-            ResourceLocation block = ore.ore_block.getRegistryName();
+            ResourceLocation block = ore.stone_ore_block.getRegistryName();
             Allomancy.LOGGER.info("Registering configured feature generation for block " + block.toString());
             ore.feature = featureFromData(ore);
             Registry.register(registry, block, ore.feature);
@@ -47,18 +51,20 @@ public class OreGenerator {
     }
 
     public static void registerGeneration(BiomeLoadingEvent event) {
-        BiomeGenerationSettingsBuilder generation = event.getGeneration();
+        var generation = event.getGeneration();
         for (OreData ore : ores) {
             if (ore.config_enabled) {
-                generation.addFeature(GenerationStage.Decoration.UNDERGROUND_ORES, ore.feature);
+                generation.addFeature(GenerationStep.Decoration.UNDERGROUND_ORES, ore.feature);
             }
         }
     }
 
     private static ConfiguredFeature<?, ?> featureFromData(OreData ore) {
+        var targetList = ImmutableList.of(OreConfiguration.target(OreConfiguration.Predicates.STONE_ORE_REPLACEABLES, ore.stone_ore_block.defaultBlockState()),
+                                          OreConfiguration.target(OreConfiguration.Predicates.DEEPSLATE_ORE_REPLACEABLES, ore.deepslate_ore_block.defaultBlockState()));
         return Feature.ORE
-                .configured(new OreFeatureConfig(OreFeatureConfig.FillerBlockType.NATURAL_STONE, ore.ore_block.defaultBlockState(), ore.vein_size))
-                .decorated(Placement.RANGE.configured(new TopSolidRangeConfig(ore.min_height, ore.min_height, ore.max_height)))
+                .configured(new OreConfiguration(targetList, ore.vein_size))
+                .rangeUniform(VerticalAnchor.absolute(ore.min_height), VerticalAnchor.absolute(ore.max_height))
                 .squared()
                 .count(ore.ores_per_chunk);
     }
@@ -69,31 +75,34 @@ public class OreGenerator {
         public final int min_height;
         public final int vein_size;
         public final int ores_per_chunk;
-        public final Block ore_block;
+        public final Block stone_ore_block;
+        public final Block deepslate_ore_block;
         public final boolean config_enabled;
         public ConfiguredFeature<?, ?> feature = null;
 
         /**
          * Construct an OreData with the given parameters
          *
-         * @param max_height     the maximum height it can generate at
-         * @param min_height     the minimum height it can generate at
-         * @param vein_size      the vein size
-         * @param ores_per_chunk number of times it can generate per chunk
-         * @param ore_block      the block to generate
-         * @param config_enabled whether or not it is enabled in the configuration file
+         * @param max_height      the maximum height it can generate at
+         * @param min_height      the minimum height it can generate at
+         * @param vein_size       the vein size
+         * @param ores_per_chunk  number of times it can generate per chunk
+         * @param stone_ore_block the block to generate
+         * @param config_enabled  whether it is enabled in the configuration file
          */
         protected OreData(ForgeConfigSpec.IntValue max_height,
                           ForgeConfigSpec.IntValue min_height,
                           ForgeConfigSpec.IntValue vein_size,
                           ForgeConfigSpec.IntValue ores_per_chunk,
-                          RegistryObject<Block> ore_block,
+                          RegistryObject<OreBlock> stone_ore_block,
+                          RegistryObject<OreBlock> ds_ore_block,
                           ForgeConfigSpec.BooleanValue config_enabled) {
             this.max_height = max_height.get();
             this.min_height = min_height.get();
             this.vein_size = vein_size.get();
             this.ores_per_chunk = ores_per_chunk.get();
-            this.ore_block = ore_block.get();
+            this.stone_ore_block = stone_ore_block.get();
+            this.deepslate_ore_block = ds_ore_block.get();
             this.config_enabled = config_enabled.get();
         }
     }
